% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{get_codes}
\alias{get_codes}
\title{Get the clinical codes in a particular code group defined
in a codes file.}
\usage{
get_codes(codes_file_path, group)
}
\description{
Get the clinical codes in a particular code group defined
in a codes file.
}
