% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rust_interface.R
\name{get_codes_in_group}
\alias{get_codes_in_group}
\title{Get a dataframe (tibble) of all the docs in a particular group
defined in a codes file.}
\usage{
get_codes_in_group(codes_file_path, group)
}
\description{
Get a dataframe (tibble) of all the docs in a particular group
defined in a codes file.
}
