% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rust_get_codes_in_group}
\alias{rust_get_codes_in_group}
\title{Get the clinical codes in a particular code group defined
in a codes file.}
\usage{
rust_get_codes_in_group(codes_file_path, group)
}
\description{
The result is a named list (intended as a dataframe) with the
columns:
\itemize{
\item name: the name of the code in the group (e.g. A01.0)
\item docs: the description of the code
}
}
\details{
TODO: figure out a good way to hand errors.
}
