% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{get_codes_in_group}
\alias{get_codes_in_group}
\title{Get the clinical codes in a particular code group defined
in a codes file.}
\usage{
get_codes_in_group(codes_file_path, group)
}
\description{
TODO: figure out a good way to hand errors
}
